Move ColumnReader factory method into ColumnReader class.
